+++ ChangeLog.md
@@ -8,6 +8,10 @@ XP Framework Core ChangeLog
 * Removed deprecated `lang.Enum::membersOf()` method - (@thekid)
 * Removed deprecated static getProxyClass() and newProxyInstance() 
   methods from the unittest.mock.MockProxyBuilder class - (@thekid)
+* Removed deprecated methods in lang.archive.Archive:
+  - addFileBytes() - replaced by addFile()
+  - add() - replaced by addBytes()
+  (@thekid)
 * Removed deprecated methods in util.Date:
   - mktime() - replaced by create()
   - fromString() - handled by regular constructor